[UPD] Ajuste na criação e remoção de arquivos temporários

+++ src/Soap/SoapBase.php
@@ -223,7 +223,7 @@ abstract class SoapBase implements SoapInterface
         $this->encriptPrivateKey = $encript;
         return $this->encriptPrivateKey;
     }
-
+   
     /**
      * Set another temporayfolder for saving certificates for SOAP utilization
      * @param string | null $folderRealPath
@@ -232,12 +232,33 @@ abstract class SoapBase implements SoapInterface
     public function setTemporaryFolder($folderRealPath = null)
     {
         if (empty($folderRealPath)) {
-            $folderRealPath = sys_get_temp_dir() . '/sped/';
+            $path = '/sped-'
+                . $this->uid()
+                .'/'
+                . $this->certificate->getCnpj()
+                . '/' ;
+            $folderRealPath = sys_get_temp_dir().$path;
+        }
+        if (substr($folderRealPath, -1) !== '/') {
+            $folderRealPath .= '/';
         }
         $this->tempdir = $folderRealPath;
         $this->setLocalFolder($folderRealPath);
     }
-
+    
+    /**
+     * Return uid from user
+     * @return string
+     */
+    protected function uid()
+    {
+        if (function_exists('posix_getuid')) {
+            return posix_getuid();
+        } else {
+            return getmyuid();
+        }
+    }
+ 
     /**
      * Set Local folder for flysystem
      * @param string $folder
@@ -438,7 +459,6 @@ abstract class SoapBase implements SoapInterface
         return $envelopeAttributes;
     }
 
-
     /**
      * Temporarily saves the certificate keys for use cURL or SoapClient
      * @return void
@@ -459,10 +479,10 @@ abstract class SoapBase implements SoapInterface
         }
         //clear dir cert
         $this->removeTemporarilyFiles();
-        $this->certsdir = $this->certificate->getCnpj() . '/certs/';
-        $this->prifile = $this->certsdir . Strings::randomString(10) . '.pem';
-        $this->pubfile = $this->certsdir . Strings::randomString(10) . '.pem';
-        $this->certfile = $this->certsdir . Strings::randomString(10) . '.pem';
+        $this->certsdir = 'certs/';
+        $this->prifile = $this->randomName();
+        $this->pubfile = $this->randomName();
+        $this->certfile = $this->randomName();
         $ret = true;
         //load private key pem
         $private = $this->certificate->privateKey;
@@ -494,6 +514,20 @@ abstract class SoapBase implements SoapInterface
             );
         }
     }
+    
+    /**
+     * Create a unique random file name
+     * @param integer $n
+     * @return string
+     */
+    protected function randomName($n = 10)
+    {
+        $name = $this->certsdir . Strings::randomString($n) . '.pem';
+        if (!$this->filesystem->has($name)) {
+            return $name;
+        }
+        $this->randomName($n+5);
+    }
 
     /**
      * Delete all files in folder
@@ -504,9 +538,23 @@ abstract class SoapBase implements SoapInterface
         if (empty($this->filesystem) || empty($this->certsdir)) {
             return;
         }
+        //remove os certificados
+        $this->filesystem->delete($this->certfile);
+        $this->filesystem->delete($this->prifile);
+        $this->filesystem->delete($this->pubfile);
+        //remove todos os arquivos antigos
         $contents = $this->filesystem->listContents($this->certsdir, true);
+        $dt = new \DateTime();
+        $tint = new \DateInterval("PT".$this->waitingTime."M");
+        $tint->invert = 1;
+        $tsLimit = $dt->add($tint)->getTimestamp();
         foreach ($contents as $item) {
-            $this->filesystem->delete($item['path']);
+            if ($item['type'] == 'file') {
+                $timestamp = $this->filesystem->getTimestamp($item['path']);
+                if ($timestamp < $tsLimit) {
+                    $this->filesystem->delete($item['path']);
+                }
+            }
         }
     }
